requirements.txt
streamlit
google-generativeai
pandas
firecrawl-py
requests